使用Layer新主题，修改properties文件为yml文件

+++ src/main/java/com/dimple/service/impl/SessionServiceImpl.java
@@ -115,9 +115,14 @@ public class SessionServiceImpl implements SessionService {
     }
 
     @Override
-    public boolean forceLogout(String sessionId) {
-        Session session = sessionDAO.readSession(sessionId);
-        sessionDAO.delete(session);
+    public boolean forceLogout(String[] sessionIds) {
+        for (String sessionId : sessionIds) {
+            Session session = sessionDAO.readSession(sessionId);
+            if (session == null) {
+                continue;
+            }
+            sessionDAO.delete(session);
+        }
         return true;
     }
 }
